Remove TypedHandler (due to issue with auto-paste unnecessary brace) Update meta (bumped version, rename desc in plugin.xml, bumped kotlin version)
