+++ build.gradle
@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlin_version = '1.2.60'
+    ext.kotlin_version = '1.2.71'
 
     repositories {
         mavenCentral()
@@ -17,7 +17,7 @@ plugins {
 }
 
 group 'org.cutejs'
-version '0.2.1'
+version '1.0.0-alpha.0'
 
 apply plugin: 'kotlin'
 apply plugin: 'org.jetbrains.grammarkit'
@@ -55,15 +55,9 @@ intellij {
 
 patchPluginXml {
     changeNotes """
-      Features:<br>
+      Changes:<br>
       <ul>
-        <li>Complete HTML markup support</li>
-        <li>JS code injection analysis (means syntax errors annotations, autocompletion, go-to-declaration etc.)</li>
-        <li>Supporting custom attributes for HTML elements: data-export-id, data-component</li>
-        <li>Brace matching for template block braces</li>
-        <li>Auto-closing template block braces</li>
-        <li>Custom icon for jst files</li>
-        <li>Custom highlighting for: type annotation blocks, export blocks, inline component blocks</li>
+        <li>Parser reborn</li>
       </ul>
       
       <a href="https://github.com/interfaced/cutejs-intellij-plugin">More info ...</a>
